implement modularization
