+++ src/index.css
@@ -7,7 +7,7 @@
   --tertiary-color: #ab99b2;
   --mono-first-color: #b3b3b3;
   --mono-second-color: #4d4d4d;
-  --mono-thirt-color: #1a1a1a;
+  --mono-third-color: #1a1a1a;
 }
 
 body {
