+++ src/pages/business/components/table/index.tsx
@@ -1,0 +1,1 @@
+export * from './Active';
