+++ src/pages/business/components/table/resolvers/DateResolver/index.tsx
@@ -1,7 +1,7 @@
 import { AtomParagraph } from 'kheiron-ui';
 import { DateType } from 'types';
 
-export const DateResolver = ({ date }: { date: any }) => {
+export const DateResolver = ({ date }: { date: string }) => {
   const __date = date as DateType;
   return <AtomParagraph>{__date}</AtomParagraph>;
 };
