--- src/pages/business/routes/Active/resolvers/StatusResolver/index.tsx
@@ -1,7 +1,0 @@
-import { AtomParagraph } from 'kheiron-ui';
-import { StatusType } from 'types';
-
-export const StatusResolver = ({ status }: { status: any }) => {
-  const __status = status as StatusType;
-  return <AtomParagraph>{__status}</AtomParagraph>;
-};
